+++ app/accounts/models.py
@@ -1,3 +1,10 @@
+from django.contrib.auth import get_user_model
 from django.db import models
 
-# Create your models here.
+
+class App(models.Model):
+    user = models.ForeignKey(
+        get_user_model(),
+        on_delete=models.CASCADE,
+    )
+    name = models.CharField(max_length=100)
