+++ app/hello_django/settings.py
@@ -37,7 +37,6 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'myapp.apps.MyappConfig',
 ]
 MIDDLEWARE = [
     'django.middleware.security.SecurityMiddleware',
