--- app/myapp/models.py
@@ -1,10 +1,0 @@
-from django.contrib.auth import get_user_model
-from django.db import models
-
-
-class App(models.Model):
-    user = models.ForeignKey(
-        get_user_model(),
-        on_delete=models.CASCADE,
-    )
-    name = models.CharField(max_length=100)
